requirements.txt
portia-grn
numpy
pandas
matplotlib
sklearn
Cython
jinja2
rpy2==3.4.5
synapseclient==2.2.2
dreamtools==1.3.0
igraph==0.9.6
